Update
Update

add similarity thresholds selection within and across individuals

Update README.md

Update README.md

+++ RADassembler/bin/CP3_Opti.pl
@@ -1,12 +1,12 @@
 #!/usr/bin/env perl
-##	Author:Yu-Long Li
+#@ Yu-Long Li
 use strict; 
 use warnings; 
 use Parallel::ForkManager;
 use Getopt::Long;
-use constant V => 171211;
+use constant V => 171216;
 
-my (@files, $file, $out_fh, $in_fh, $in_tags, $sub_lst);
+my (@files, $file, $out_fh, $in_fh, $in_tags, $sub_lst, $sec_asmb);
 my $in_path;
 my $out_path;
 my $loci;
@@ -24,7 +24,8 @@ my $usage =
 	-t number of threads [1]
 	-c collect the assembly results for the second reads [0]
 	-p parameters parsed to CAP3
-	-f sub file list for assembly 
+	-f sub file list for assembly
+	-2 continue to run the second assembly
 	-h help
 ';
 
@@ -35,6 +36,7 @@ GetOptions ("par=s" 		=> \$cmd,
 			"threads=i" 	=> \$num_threads,
 			"collect:1" 	=> \$collect,
 			"file:s" 	    => \$sub_lst,
+			"2:1"			=> \$sec_asmb,
 			"help:1"		=> \$help
 			)
 			or die ("Error in command line arguments\n");
@@ -54,8 +56,6 @@ if (! -e "$out_path/assembly_1st") {
 
 `find $in_path/ -name "*cap*" |xargs rm -rf`;
 
-
-print "Starting the first run...\n";
 unless($cmd) {
 	#Parameters for short reads.
 	$cmd = '-r 0 -i 30 -j 31 -o 18 -s 300 -p 85';
@@ -106,30 +106,30 @@ if ($sub_lst) {
 
 ##	multi-threading part for the 1st assembly;
 my $thread_m = new Parallel::ForkManager($num_threads);
-my $total = @files;
-my $i;
-foreach $file (@files){
+unless ($sec_asmb) {
 	
-	$i++;
-	$thread_m->start and next;
+	print "Starting the first run...\n";
+	my $total = @files;
+	my $i;
+	foreach $file (@files){
 	
-	run_assembly($file, $in_path, "$out_path/assembly_1st", 1, $cmd);
-	print STDERR "Assembling locus $i of $total \r";
-	
-	$thread_m->finish;
-	
-	
-}
+		$i++;
+		$thread_m->start and next;
+		run_assembly($file, $in_path, "$out_path/assembly_1st", 1, $cmd);
+		
+		print STDERR "Assembling locus $i of $total \r";
+		$thread_m->finish;
+		
+	}
 
-$thread_m->wait_all_children;
-`find $in_path/ -name "*cap*" |xargs rm -rf`;
+	$thread_m->wait_all_children;
+	`find $in_path/ -name "*cap*" |xargs rm -rf`;
+}
 
 ## 2nd assembly;
 print "\nStarting the second run...\n";
-#######
 
 `echo -e "cap3 $cmd" > $out_path/log/assembly.2par`;
-#######
 
 if (! -e "$out_path/assembly_2nd") {
 	
@@ -137,25 +137,21 @@ if (! -e "$out_path/assembly_2nd") {
 	
 }
 
-
 get_flist("$out_path/assembly_1st/", $out_path, "reads_2nd", "fa");
 
+$thread_m = new Parallel::ForkManager($num_threads*2); # the second run could be 2X faster.
 foreach $file (@files){
 	
-	
-	
-	#print "start one threads" . "\n";
-
 	$thread_m->start and next;
 	
 	run_assembly($file, "$out_path/assembly_1st", "$out_path/assembly_2nd", 2, $cmd);
 	
 	$thread_m->finish;
 	
-	
 }
 
 $thread_m->wait_all_children;
+
 `find $out_path/assembly_1st/ -name "*cap*" |xargs rm -rf`;
 
 ##collect the final results, and delet useless files;
@@ -165,8 +161,6 @@ get_flist("$out_path/assembly_2nd", $out_path, "2nd_assembled","fa");	# get the 
 
 parse_fasta(\@files, "$out_path/assembly_2nd", "$out_path", "collected_final.fa");
 
-
-
 ##	if collect contigs for reads1;
 if ($collect) {
 	
@@ -174,8 +168,8 @@ if ($collect) {
 	my $out_fa = $out_path;
 	my $name_fa = "collected_reads1.fa";
 	
-	
 	get_flist("$out_path/assembly_1st",$out_path,"collect_reads1","fa");
+	
 	parse_fasta(\@files, $in_fa, $out_fa, $name_fa);
 	
 	`sed -i /Consensus/d  $out_fa/$name_fa`;
@@ -193,19 +187,21 @@ sub parse_fasta {
 	foreach $file (@$files) {
 		open ($in_fh, "<$in_fa/$file") or die "$!";
 		while (<$in_fh>) {
-		
-		chomp if /Consensus/;
-		$_ =~ s/^>/>$file\./;
-		print $out_fh $_;
-		
+			chomp if /Consensus/;
+			my $locus = substr($file, 0, -3);
+			if (/^>/) {
+				print $out_fh "\n" if $. > 1; # a new start of fasta.
+				$_ =~ s/^>/>$locus\./;
+			} else {
+				$_ =~ s/\n$//;
+			}
+			print $out_fh $_;
 		}
-		
+		print $out_fh "\n";
 		close $in_fh;
 	}
 	
-	#print $out_fh "\n";	#the last line;
 	close $out_fh;
-	
 	
 }
 
@@ -223,7 +219,13 @@ sub run_assembly {
             my $read1 = `grep '_Consensus' $in_path/$file.cap.singlets`;
             if ($read1) {
                 # read1 is not assembled, link it.
-                `cat $in_path/$file.cap.singlets >> $out_path/$file`;
+				my $locus = substr($file, 0, -3);
+				my $seq   = '>'.$locus.'_Consensus'."\n";
+				   $seq  .= $seqs{$locus};
+				open(my $out_fh, ">>$out_path/$file") or die "$!";
+				print $out_fh $seq;
+				close $out_fh;
+                #`echo "$seq" >> $out_path/$file`;
                 link_fasta($out_path, $out_path, $file);
             }
 		}
@@ -251,7 +253,6 @@ sub get_flist {
 	my ($in_path, $out_path, $name, $type) = @_;
 	
 	`mkdir -p $out_path/log/`;
-	
 	
 	undef (@files);
 	opendir (D, $in_path) or die "$!";
@@ -305,8 +306,12 @@ sub link_fasta {
     if ($num == 1) {
         $seq = $buf;
     } else {
-        $id = '>Contig1_10N_read1';
-        $seq = $ctg . 'N'x10 . $read1;
+        $id = '>ctg_10N_rd1';
+		if ($ctg =~ /NNNNNNNNNN$/) {
+			$seq = $ctg . $read1;
+		} else {
+			$seq = $ctg . 'N'x10 . $read1;
+		}
     }
     open (my $out_fh, ">$out_path/$file");
     print $out_fh $id, "\n",
